Extract PullIndexJournal from ImportClient
Move the pull/index.wal writer and the merge which publishes its records
into the remote and local indexes out of ImportClient into a dedicated
PullIndexJournal class. ImportClient keeps the cursor and state ordering;
the journal only guarantees that flushed records are durable before each
cursor checkpoint.
