+++ packages/reprint-client/src/import.php
@@ -25,6 +25,8 @@ use Reprint\Importer\Tuning\AdaptiveTuner;
 
 use function Reprint\Importer\apply_curl_ca_bundle;
 use function Reprint\Importer\apply_curl_proxy_from_environment;
+use function Reprint\Importer\parse_index_line;
+use function Reprint\Importer\read_remote_index_entry;
 use function Reprint\Importer\register_sqlite_function;
 use function Reprint\Importer\resolve_sqlite_integration_path;
 use function Reprint\Importer\resolve_sqlite_integration_plugin_path;
@@ -100,6 +102,10 @@ require_once __DIR__ . '/lib/import/load.php';
 
 // High-level pull commands — orchestrate lower-level commands into pipelines
 require_once __DIR__ . '/lib/pull/class-pull.php';
+
+// Pull index journal — the WAL for completed files-pull mutations.
+require_once __DIR__ . '/lib/pull/remote-index-functions.php';
+require_once __DIR__ . '/lib/pull/class-pull-index-journal.php';
 
 /**
  * The wire-protocol version this importer speaks.
@@ -209,8 +215,8 @@ class ImportClient
      */
     private $pull_index_wal_path;
 
-    /** @var resource|null Open file handle for $pull_index_wal_path while writing. */
-    private $pull_index_wal_handle;
+    /** @var PullIndexJournal Owns pull/index.wal and applies its records to both indexes. */
+    private $pull_index_journal;
 
     /**
      * @var string Next remote index file pull/remote-index.next.jsonl, including
@@ -496,6 +502,14 @@ class ImportClient
         }
         $this->filesystem_root = $resolved_local_filesystem_root;
 
+        $this->pull_index_journal = new PullIndexJournal(
+            $this,
+            $this->pull_index_wal_path,
+            $this->remote_index_file,
+            $this->local_index_file,
+            $this->filesystem_root
+        );
+
         $this->state = new PullState();
     }
 
@@ -517,195 +531,6 @@ class ImportClient
         }
         fclose($remote_index_file_handle);
         return $remote_index_entry_count;
-    }
-
-    /**
-     * Upsert an entry in the remote index.
-     */
-    private function upsert_remote_index_entry(
-        string $remote_absolute_path,
-        int $remote_path_ctime,
-        int $remote_path_size,
-        string $remote_path_type
-    ): void {
-        $this->write_pull_index_wal_record([
-            "op" => "+",
-            "remote_absolute_path_b64" => base64_encode($remote_absolute_path),
-            "remote_path_ctime" => $remote_path_ctime,
-            "remote_path_size" => $remote_path_size,
-            "remote_path_type" => $remote_path_type,
-        ]);
-    }
-
-    /** Appends a completed local deletion to the pull index WAL. */
-    private function wal_append_successful_deletion(
-        string $remote_absolute_path,
-        string $local_absolute_path
-    ): void {
-        $pull_index_wal_record = [
-            "op" => "-",
-            "remote_absolute_path_b64" => base64_encode($remote_absolute_path),
-        ];
-        $local_relative_path = $this->local_relative_path_from_local_absolute_path(
-            $local_absolute_path
-        );
-        if ($local_relative_path !== null) {
-            $pull_index_wal_record["local_relative_path_b64"] =
-                base64_encode($local_relative_path);
-        }
-        $this->write_pull_index_wal_record($pull_index_wal_record);
-    }
-
-    /** Invalidates remote state which this pull did not account for locally. */
-    private function wal_append_remote_index_invalidation(string $remote_absolute_path): void
-    {
-        $this->write_pull_index_wal_record([
-            "op" => "-",
-            "remote_absolute_path_b64" => base64_encode($remote_absolute_path),
-        ]);
-    }
-
-    /**
-     * Appends a completed local upsert to the pull index WAL.
-     *
-     * Files-pull calls this only after the local absolute path contains the
-     * pulled file, symlink, or empty directory. For example:
-     *
-     *     filesystem root:      /var/www
-     *     remote absolute path: /srv/site/file.txt
-     *     local absolute path:  /var/www/file.txt
-     *
-     * Applying the WAL produces decoded entries such as:
-     *
-     *     remote index: /srv/site/file.txt  file, size 4, ctime 10
-     *     local index:  file.txt            file, size 4, ctime 12
-     *
-     * The remote index records the remote state files-pull accounted for. The
-     * local index records the resulting local path type, size, and ctime.
-     * Without that local index entry, files-diff and PushPlan would compare
-     * the pulled path with the older local index and select it as a local
-     * change. A local absolute path outside the filesystem root, or a
-     * default-skipped path, has no local index entry.
-     */
-    private function record_pulled_path(
-        string $remote_absolute_path,
-        string $local_absolute_path,
-        int $remote_path_ctime,
-        int $remote_path_size,
-        string $remote_path_type
-    ): void {
-        $pull_index_wal_record = [
-            "op" => "+",
-            "remote_absolute_path_b64" => base64_encode($remote_absolute_path),
-            "remote_path_ctime" => $remote_path_ctime,
-            "remote_path_size" => $remote_path_size,
-            "remote_path_type" => $remote_path_type,
-        ];
-        $local_relative_path = $this->local_relative_path_from_local_absolute_path(
-            $local_absolute_path
-        );
-        if ($local_relative_path !== null) {
-            clearstatcache(true, $local_absolute_path);
-            $local_path_stat = lstat($local_absolute_path);
-            if ($local_path_stat === false) {
-                // phpcs:disable WordPress.Security.EscapeOutput.ExceptionNotEscaped -- CLI filesystem path, never HTML output.
-                throw new RuntimeException(
-                    "Failed to inspect the pulled local absolute path: {$local_absolute_path}."
-                );
-                // phpcs:enable WordPress.Security.EscapeOutput.ExceptionNotEscaped
-            }
-            $local_file_type_bits = $local_path_stat["mode"] & 0170000;
-            if ($local_file_type_bits === 0120000) {
-                $local_path_type = "link";
-            } elseif ($local_file_type_bits === 0040000) {
-                $local_path_type = "dir";
-            } elseif ($local_file_type_bits === 0100000) {
-                $local_path_type = "file";
-            } else {
-                // phpcs:disable WordPress.Security.EscapeOutput.ExceptionNotEscaped -- CLI filesystem path, never HTML output.
-                throw new RuntimeException(
-                    "The pulled local absolute path has an unsupported type: {$local_absolute_path}."
-                );
-                // phpcs:enable WordPress.Security.EscapeOutput.ExceptionNotEscaped
-            }
-            $pull_index_wal_record["local_relative_path_b64"] =
-                base64_encode($local_relative_path);
-            $pull_index_wal_record["local_path_ctime"] = (int) $local_path_stat["ctime"];
-            $pull_index_wal_record["local_path_size"] =
-                $local_path_type === "dir" ? 0 : (int) $local_path_stat["size"];
-            $pull_index_wal_record["local_path_type"] = $local_path_type;
-        }
-        $this->write_pull_index_wal_record($pull_index_wal_record);
-    }
-
-    /** Returns the local relative path stored in the local index. */
-    private function local_relative_path_from_local_absolute_path(
-        string $local_absolute_path
-    ): ?string {
-        $local_relative_path = relative_path_under(
-            $local_absolute_path,
-            $this->filesystem_root
-        );
-        if (
-            $local_relative_path === null
-            || $local_relative_path === ""
-        ) {
-            return null;
-        }
-        return FileIndexProcessor::path_is_default_skipped(
-            $local_relative_path
-        )
-            ? null
-            : $local_relative_path;
-    }
-
-    /**
-     * Appends one complete record to the pull index WAL.
-     *
-     * @param array $pull_index_wal_record {
-     *     One completed pull mutation, with local fields when files-pull
-     *     changed a non-skipped path beneath the filesystem root.
-     *
-     *     @type string $op                       `+` upsert or `-` deletion.
-     *     @type string $remote_absolute_path_b64 Base64 remote absolute path.
-     *     @type int    $remote_path_ctime        Remote ctime for `+`.
-     *     @type int    $remote_path_size         Remote size for `+`.
-     *     @type string $remote_path_type         Remote type for `+`.
-     *     @type string $local_relative_path_b64  Base64 local relative path
-     *                                             when the completed mutation
-     *                                             belongs in the local index.
-     *     @type int    $local_path_ctime         Local ctime for a local `+`.
-     *     @type int    $local_path_size          Local size for a local `+`.
-     *     @type string $local_path_type          Local type for a local `+`.
-     * }
-     */
-    private function write_pull_index_wal_record(
-        array $pull_index_wal_record
-    ): void
-    {
-        if (!$this->pull_index_wal_handle) {
-            $this->open_pull_index_wal();
-        }
-        $pull_index_wal_json_line = json_encode(
-            $pull_index_wal_record,
-            JSON_UNESCAPED_SLASHES | JSON_THROW_ON_ERROR
-        ) . "\n";
-        if (
-            fwrite($this->pull_index_wal_handle, $pull_index_wal_json_line)
-            !== strlen($pull_index_wal_json_line)
-        ) {
-            throw new RuntimeException(
-                "Failed to write to the pull index WAL (disk full?)."
-            );
-        }
-    }
-
-    /** Replays a pull index WAL left by an interrupted batch. */
-    private function replay_pull_index_wal(): void
-    {
-        if (is_file($this->pull_index_wal_path)) {
-            $this->apply_pull_index_wal();
-        }
     }
 
     /**
@@ -2440,10 +2265,9 @@ class ImportClient
             true,
         );
         // Replay the pull index WAL before clearing the cursor which made its records durable.
-        $this->replay_pull_index_wal();
-        $this->remove_pull_index_wal();
+        $this->pull_index_journal->apply_pending();
+        $this->pull_index_journal->remove_empty_marker();
         $this->reset_state();
-        $this->pull_index_wal_handle = null;
 
         if (file_exists($this->next_remote_index_file)) {
             @unlink($this->next_remote_index_file);
@@ -3097,13 +2921,13 @@ class ImportClient
             $current_status !== null &&
             $current_status !== "complete";
 
-        $this->replay_pull_index_wal();
+        $this->pull_index_journal->apply_pending();
         $this->assert_files_pull_path_selection_unchanged_while_resuming($has_progress);
         $this->assert_local_followed_symlinks_root_unchanged();
 
         // Already completed.
         if ($current_status === "complete") {
-            $this->remove_pull_index_wal();
+            $this->pull_index_journal->remove_empty_marker();
             $remote_index_entry_count = $this->remote_index_entry_count();
             $this->progress->clear_progress_line();
 
@@ -3182,7 +3006,7 @@ class ImportClient
 
             // The marker blocks files-diff and files-push before the first
             // pull checkpoint can make this lifecycle resumable.
-            $this->open_pull_index_wal();
+            $this->pull_index_journal->open();
             $this->get_state()->active_resumable_command->command_name = "files-pull";
             $this->get_state()->active_resumable_command->completion_state = "in_progress";
             $this->get_state()->active_resumable_command->current_stage = "index";
@@ -3234,7 +3058,7 @@ class ImportClient
         $this->get_state()->active_resumable_command->completion_state = "in_progress";
         $this->save_state();
 
-        $this->open_pull_index_wal();
+        $this->pull_index_journal->open();
         $stage = $this->get_state()->active_resumable_command->current_stage ?? "index";
 
         if ($stage === "index") {
@@ -3331,12 +3155,12 @@ class ImportClient
         if ($this->follow_symlinks) {
             $this->recreate_intermediate_symlinks();
         }
-        $this->apply_pull_index_wal();
+        $this->pull_index_journal->apply_pending();
 
         $this->ensure_local_index_exists();
         $this->get_state()->active_resumable_command->completion_state = "complete";
         $this->save_state();
-        $this->remove_pull_index_wal();
+        $this->pull_index_journal->remove_empty_marker();
 
         $this->progress->clear_progress_line();
         $remote_index_entry_count = $this->remote_index_entry_count();
@@ -3831,12 +3655,12 @@ class ImportClient
 
             if (@symlink($symlink_target, $local_absolute_path)) {
                 $created++;
-                $this->record_pulled_path(
+                $this->pull_index_journal->record_remote_upsert(
                     $remote_absolute_path,
-                    $local_absolute_path,
                     (int) ($next_remote_index_entry["ctime"] ?? 0),
                     (int) ($next_remote_index_entry["size"] ?? 0),
-                    "link"
+                    "link",
+                    $local_absolute_path
                 );
                 $this->audit_log(
                     "INTERMEDIATE SYMLINK: {$remote_absolute_path} -> {$symlink_target}",
@@ -4114,7 +3938,7 @@ class ImportClient
             $next_remote_index_file_handle = fopen($next_remote_index_file, "r");
             if ($next_remote_index_file_handle) {
                 while (($next_remote_index_json_line = fgets($next_remote_index_file_handle)) !== false) {
-                    $next_remote_index_entry = $this->parse_index_line($next_remote_index_json_line);
+                    $next_remote_index_entry = parse_index_line($next_remote_index_json_line);
                     if ($next_remote_index_entry === null) {
                         continue;
                     }
@@ -6401,12 +6225,7 @@ class ImportClient
                         $this->get_state()->current_file = null;
                         $this->get_state()->current_file_bytes = null;
                     }
-                    if (
-                        $this->pull_index_wal_handle
-                        && !fflush($this->pull_index_wal_handle)
-                    ) {
-                        throw new RuntimeException('Failed to flush the pull index WAL.');
-                    }
+                    $this->pull_index_journal->flush();
                     $this->get_state()->fetch->cursor = $cursor;
                     $this->save_state();
                     $chunks_since_save = 0;
@@ -6441,7 +6260,7 @@ class ImportClient
                 fclose($context->file_handle);
                 $context->file_handle = null;
             }
-            $this->apply_pull_index_wal();
+            $this->pull_index_journal->apply_pending();
             $this->get_state()->active_resumable_command->completion_state = "partial";
             $this->save_state();
             return false;
@@ -6455,7 +6274,7 @@ class ImportClient
             $context->response_stats ?? [],
         );
         $this->get_state()->fetch->cursor = $cursor;
-        $this->apply_pull_index_wal();
+        $this->pull_index_journal->apply_pending();
         // Update file tracking: track in-progress file, or clear if complete/no active file
         if ($context->file_handle && $context->file_path) {
             if (!fflush($context->file_handle)) {
@@ -6768,7 +6587,7 @@ class ImportClient
             ? fopen($this->remote_index_file, "r")
             : null;
         $remote_index_entry =
-            $this->read_remote_index_entry($remote_index_file_handle);
+            read_remote_index_entry($remote_index_file_handle);
         if ($last_consumed_remote_index_entry_path) {
             while (
                 $remote_index_entry !== null &&
@@ -6778,10 +6597,10 @@ class ImportClient
                 ) <= 0
             ) {
                 $remote_index_entry =
-                    $this->read_remote_index_entry($remote_index_file_handle);
+                    read_remote_index_entry($remote_index_file_handle);
             }
         }
-        $this->open_pull_index_wal();
+        $this->pull_index_journal->open();
         $next_remote_index_entries_processed = 0;
 
         while (($next_remote_index_json_line = fgets($next_remote_index_file_handle)) !== false) {
@@ -6794,7 +6613,7 @@ class ImportClient
             }
 
             $next_remote_index_byte_offset = ftell($next_remote_index_file_handle);
-            $next_remote_index_entry = $this->parse_index_line($next_remote_index_json_line);
+            $next_remote_index_entry = parse_index_line($next_remote_index_json_line);
             if (!$next_remote_index_entry) {
                 continue;
             }
@@ -6816,11 +6635,11 @@ class ImportClient
                         $remote_deletion_root
                     );
                     if ($local_absolute_path === null) {
-                        $this->wal_append_remote_index_invalidation(
+                        $this->pull_index_journal->record_remote_invalidation(
                             $missing_remote_index_entry_path
                         );
                     } else {
-                        $this->wal_append_successful_deletion(
+                        $this->pull_index_journal->record_successful_deletion(
                             $missing_remote_index_entry_path,
                             $local_absolute_path
                         );
@@ -6829,7 +6648,7 @@ class ImportClient
                 $last_consumed_remote_index_entry_path =
                     $remote_index_entry["path"];
                 $remote_index_entry =
-                    $this->read_remote_index_entry($remote_index_file_handle);
+                    read_remote_index_entry($remote_index_file_handle);
             }
 
             if (
@@ -6854,7 +6673,7 @@ class ImportClient
                 $last_consumed_remote_index_entry_path =
                     $remote_index_entry["path"];
                 $remote_index_entry =
-                    $this->read_remote_index_entry($remote_index_file_handle);
+                    read_remote_index_entry($remote_index_file_handle);
             } elseif (
                 $this->is_selected_for_pulling($next_remote_index_entry["path"], true) &&
                 (
@@ -6886,12 +6705,7 @@ class ImportClient
                     $last_consumed_remote_index_entry_path;
                 $this->get_state()->diff->last_processed_next_remote_index_entry_path =
                     $last_processed_next_remote_index_entry_path;
-                if (
-                    $this->pull_index_wal_handle
-                    && !fflush($this->pull_index_wal_handle)
-                ) {
-                    throw new RuntimeException('Failed to flush the pull index WAL.');
-                }
+                $this->pull_index_journal->flush();
                 $this->save_state();
                 $this->progress->tick_spinner();
             }
@@ -6909,11 +6723,11 @@ class ImportClient
                     $remote_deletion_root
                 );
                 if ($local_absolute_path === null) {
-                    $this->wal_append_remote_index_invalidation(
+                    $this->pull_index_journal->record_remote_invalidation(
                         $missing_remote_index_entry_path
                     );
                 } else {
-                    $this->wal_append_successful_deletion(
+                    $this->pull_index_journal->record_successful_deletion(
                         $missing_remote_index_entry_path,
                         $local_absolute_path
                     );
@@ -6922,7 +6736,7 @@ class ImportClient
             $last_consumed_remote_index_entry_path =
                 $remote_index_entry["path"];
             $remote_index_entry =
-                $this->read_remote_index_entry($remote_index_file_handle);
+                read_remote_index_entry($remote_index_file_handle);
         }
 
         if ($remote_index_file_handle) {
@@ -6936,7 +6750,7 @@ class ImportClient
             $last_consumed_remote_index_entry_path;
         $this->get_state()->diff->last_processed_next_remote_index_entry_path =
             $last_processed_next_remote_index_entry_path;
-        $this->apply_pull_index_wal();
+        $this->pull_index_journal->apply_pending();
         $this->save_state();
 
         return !$this->shutdown_requested;
@@ -7389,395 +7203,6 @@ class ImportClient
         }
 
         return true === @unlink($local_absolute_path);
-    }
-
-    /**
-     * Parse one JSON index line into an array.
-     */
-    private function parse_index_line(string $line): ?array
-    {
-        $line = trim($line);
-        if ($line === "") {
-            return null;
-        }
-        $data = json_decode($line, true);
-        if (!is_array($data)) {
-            throw new RuntimeException("Invalid index line format");
-        }
-        $path_encoded = $data["path"] ?? "";
-        if (!is_string($path_encoded) || $path_encoded === "") {
-            throw new RuntimeException("Invalid index path");
-        }
-        $path = base64_decode($path_encoded, true);
-        if ($path === "" || $path === false) {
-            throw new RuntimeException("Invalid index path (base64 decode failed)");
-        }
-        assert_valid_path($path, "index path");
-        return [
-            "path" => $path,
-            "ctime" => (int) ($data["ctime"] ?? 0),
-            "size" => (int) ($data["size"] ?? 0),
-            "type" => (string) ($data["type"] ?? "file"),
-        ];
-    }
-
-    /** Opens the current pull index WAL for append. */
-    private function open_pull_index_wal(): void
-    {
-        if ($this->pull_index_wal_handle) {
-            return;
-        }
-        $pull_index_wal_is_new = !is_file($this->pull_index_wal_path);
-        $this->pull_index_wal_handle = fopen($this->pull_index_wal_path, "a");
-        if (!$this->pull_index_wal_handle) {
-            throw new RuntimeException("Failed to open the pull index WAL.");
-        }
-        if ($pull_index_wal_is_new) {
-            $this->audit_log(
-                "FILE CREATE | {$this->pull_index_wal_path} | pull index WAL",
-            );
-        }
-    }
-
-    /** Applies the pull index WAL to the remote index and then the local index. */
-    private function apply_pull_index_wal(): void
-    {
-        if ($this->pull_index_wal_handle) {
-            $pull_index_wal_closed = fclose($this->pull_index_wal_handle);
-            $this->pull_index_wal_handle = null;
-            if (!$pull_index_wal_closed) {
-                throw new RuntimeException("Failed to flush the pull index WAL.");
-            }
-        }
-        clearstatcache(true, $this->pull_index_wal_path);
-        if (
-            !is_file($this->pull_index_wal_path)
-            || filesize($this->pull_index_wal_path) === 0
-        ) {
-            return;
-        }
-
-        $remote_index_replacement_file = $this->remote_index_file . ".new";
-
-        $this->audit_log(
-            "INDEX MERGE START | merging pull index WAL into {$this->remote_index_file}",
-        );
-
-        $remote_index_file_handle = file_exists($this->remote_index_file)
-            ? fopen($this->remote_index_file, "r")
-            : null;
-        $pull_index_wal_file_handle = fopen($this->pull_index_wal_path, "r");
-        $remote_index_replacement_file_handle = fopen($remote_index_replacement_file, "w");
-
-        if (!$pull_index_wal_file_handle || !$remote_index_replacement_file_handle) {
-            throw new RuntimeException("Failed to merge remote index updates.");
-        }
-
-        $write_remote_index_entry = function ($remote_index_destination_file_handle, array $remote_index_entry_to_write): void {
-            $remote_index_json_line = json_encode(
-                [
-                    "path" => base64_encode($remote_index_entry_to_write["path"]),
-                    "ctime" => (int) $remote_index_entry_to_write["ctime"],
-                    "size" => (int) $remote_index_entry_to_write["size"],
-                    "type" => (string) $remote_index_entry_to_write["type"],
-                ],
-                JSON_UNESCAPED_SLASHES,
-            );
-            if ($remote_index_json_line !== false) {
-                fwrite($remote_index_destination_file_handle, $remote_index_json_line . "\n");
-            }
-        };
-
-        $remote_index_entry = $this->read_remote_index_entry($remote_index_file_handle);
-        $remote_index_update_lookahead = null;
-        $remote_index_update = $this->read_remote_index_update(
-            $pull_index_wal_file_handle,
-            $remote_index_update_lookahead
-        );
-        $last_written_remote_index_entry_path = null;
-
-        while ($remote_index_entry !== null || $remote_index_update !== null) {
-            if ($remote_index_update === null) {
-                if ($last_written_remote_index_entry_path !== $remote_index_entry["path"]) {
-                    $write_remote_index_entry($remote_index_replacement_file_handle, $remote_index_entry);
-                    $last_written_remote_index_entry_path = $remote_index_entry["path"];
-                }
-                $remote_index_entry = $this->read_remote_index_entry($remote_index_file_handle);
-                continue;
-            }
-
-            if ($remote_index_entry === null) {
-                if (
-                    !$remote_index_update["delete"] &&
-                    $last_written_remote_index_entry_path !== $remote_index_update["path"]
-                ) {
-                    $write_remote_index_entry($remote_index_replacement_file_handle, $remote_index_update);
-                    $last_written_remote_index_entry_path = $remote_index_update["path"];
-                }
-                $remote_index_update = $this->read_remote_index_update(
-                    $pull_index_wal_file_handle,
-                    $remote_index_update_lookahead
-                );
-                continue;
-            }
-
-            $remote_index_entry_path_comparison = strcmp($remote_index_entry["path"], $remote_index_update["path"]);
-            if ($remote_index_entry_path_comparison === 0) {
-                if (
-                    !$remote_index_update["delete"] &&
-                    $last_written_remote_index_entry_path !== $remote_index_update["path"]
-                ) {
-                    $write_remote_index_entry($remote_index_replacement_file_handle, $remote_index_update);
-                    $last_written_remote_index_entry_path = $remote_index_update["path"];
-                }
-                $remote_index_entry = $this->read_remote_index_entry($remote_index_file_handle);
-                $remote_index_update = $this->read_remote_index_update(
-                    $pull_index_wal_file_handle,
-                    $remote_index_update_lookahead
-                );
-            } elseif ($remote_index_entry_path_comparison < 0) {
-                if ($last_written_remote_index_entry_path !== $remote_index_entry["path"]) {
-                    $write_remote_index_entry($remote_index_replacement_file_handle, $remote_index_entry);
-                    $last_written_remote_index_entry_path = $remote_index_entry["path"];
-                }
-                $remote_index_entry = $this->read_remote_index_entry($remote_index_file_handle);
-            } else {
-                if (
-                    !$remote_index_update["delete"] &&
-                    $last_written_remote_index_entry_path !== $remote_index_update["path"]
-                ) {
-                    $write_remote_index_entry($remote_index_replacement_file_handle, $remote_index_update);
-                    $last_written_remote_index_entry_path = $remote_index_update["path"];
-                }
-                $remote_index_update = $this->read_remote_index_update(
-                    $pull_index_wal_file_handle,
-                    $remote_index_update_lookahead
-                );
-            }
-        }
-
-        if ($remote_index_file_handle) {
-            fclose($remote_index_file_handle);
-        }
-        fclose($pull_index_wal_file_handle);
-        fclose($remote_index_replacement_file_handle);
-
-        if (!rename($remote_index_replacement_file, $this->remote_index_file)) {
-            throw new RuntimeException("Failed to replace the remote index file.");
-        }
-        $this->audit_log("INDEX MERGE COMPLETE | {$this->remote_index_file} updated");
-
-        /*
-         * Rebuild the sorted local index updates from the pull index WAL. This
-         * temporary file is disposable: the WAL remains until both index
-         * replacements finish, so resume can discard a partial file and replay
-         * the batch. The WAL is in completion order, while the local index
-         * merge requires local relative path byte order. Records without a
-         * local relative path update only the remote index. An unterminated
-         * final record is repeated from the preceding durable cursor when
-         * files-pull resumes.
-         */
-        $local_index_updates_path = $this->pull_index_wal_path . ".local";
-        $pull_index_wal_file_handle = fopen($this->pull_index_wal_path, "r");
-        $local_index_updates_handle = fopen($local_index_updates_path, "w");
-        if (!$pull_index_wal_file_handle || !$local_index_updates_handle) {
-            throw new RuntimeException("Failed to prepare the local index updates.");
-        }
-
-        $local_index_updates_written = 0;
-        while (( $pull_index_wal_json_line = fgets($pull_index_wal_file_handle) ) !== false) {
-            if (
-                substr($pull_index_wal_json_line, -1) !== "\n"
-                && feof($pull_index_wal_file_handle)
-            ) {
-                break;
-            }
-            $pull_index_wal_record = json_decode($pull_index_wal_json_line, true);
-            if (!is_array($pull_index_wal_record)) {
-                throw new RuntimeException("Invalid pull index WAL line format.");
-            }
-            if (!array_key_exists("local_relative_path_b64", $pull_index_wal_record)) {
-                continue;
-            }
-            $local_index_update = [
-                "op" => $pull_index_wal_record["op"],
-                "path" => $pull_index_wal_record["local_relative_path_b64"],
-            ];
-            if ($pull_index_wal_record["op"] === "+") {
-                $local_index_update += [
-                    "ctime" => $pull_index_wal_record["local_path_ctime"],
-                    "size" => $pull_index_wal_record["local_path_size"],
-                    "type" => $pull_index_wal_record["local_path_type"],
-                ];
-            }
-            write_local_index_update(
-                $local_index_updates_handle,
-                $local_index_update
-            );
-            ++$local_index_updates_written;
-        }
-        fclose($pull_index_wal_file_handle);
-        fclose($local_index_updates_handle);
-
-        if ($local_index_updates_written > 0) {
-            sort_index_file($local_index_updates_path);
-            merge_local_index_mutations(
-                $this->local_index_file,
-                $local_index_updates_path
-            );
-        }
-        @unlink($local_index_updates_path);
-
-        if (file_put_contents($this->pull_index_wal_path, "") === false) {
-            throw new RuntimeException(
-                "Failed to clear the applied pull index WAL."
-            );
-        }
-        $this->audit_log(
-            "FILE TRUNCATE | {$this->pull_index_wal_path} | pull index WAL batch applied"
-        );
-    }
-
-    /** Removes the pull index WAL marker after files-pull completes or is aborted. */
-    private function remove_pull_index_wal(): void
-    {
-        if (is_resource($this->pull_index_wal_handle)) {
-            if (!fclose($this->pull_index_wal_handle)) {
-                throw new RuntimeException("Failed to flush the pull index WAL.");
-            }
-            $this->pull_index_wal_handle = null;
-        }
-        clearstatcache(true, $this->pull_index_wal_path);
-        if (
-            is_file($this->pull_index_wal_path)
-            && filesize($this->pull_index_wal_path) > 0
-        ) {
-            throw new RuntimeException(
-                "Cannot remove an unapplied pull index WAL."
-            );
-        }
-        if (
-            is_file($this->pull_index_wal_path)
-            && !unlink($this->pull_index_wal_path)
-        ) {
-            throw new RuntimeException("Failed to remove the pull index WAL.");
-        }
-    }
-
-    /** Reads one entry from the remote index. */
-    private function read_remote_index_entry($remote_index_file_handle): ?array
-    {
-        if (!$remote_index_file_handle) {
-            return null;
-        }
-        while (($remote_index_json_line = fgets($remote_index_file_handle)) !== false) {
-            $remote_index_entry = $this->parse_index_line($remote_index_json_line);
-            if ($remote_index_entry !== null) {
-                return $remote_index_entry;
-            }
-        }
-        return null;
-    }
-
-    /** Reads one raw remote index projection from the pull index WAL. */
-    private function read_raw_remote_index_update(
-        $pull_index_wal_file_handle
-    ): ?array {
-        if (!$pull_index_wal_file_handle) {
-            return null;
-        }
-        while (( $pull_index_wal_json_line = fgets($pull_index_wal_file_handle) ) !== false) {
-            if (substr($pull_index_wal_json_line, -1) !== "\n" && feof($pull_index_wal_file_handle)) {
-                return null;
-            }
-            $pull_index_wal_json_line = trim($pull_index_wal_json_line);
-            if ($pull_index_wal_json_line === "") {
-                continue;
-            }
-            $pull_index_wal_record = json_decode($pull_index_wal_json_line, true);
-            if (!is_array($pull_index_wal_record)) {
-                throw new RuntimeException("Invalid pull index WAL line format.");
-            }
-            $pull_index_wal_operation = $pull_index_wal_record["op"] ?? null;
-            $remote_absolute_path_base64 =
-                $pull_index_wal_record["remote_absolute_path_b64"] ?? null;
-            if (
-                !is_string($remote_absolute_path_base64)
-                || $remote_absolute_path_base64 === ""
-            ) {
-                throw new RuntimeException(
-                    "Invalid pull index WAL remote absolute path."
-                );
-            }
-            $remote_absolute_path = base64_decode($remote_absolute_path_base64, true);
-            if ($remote_absolute_path === false || $remote_absolute_path === "") {
-                throw new RuntimeException(
-                    "Invalid pull index WAL remote absolute path (base64 decode failed)."
-                );
-            }
-            if ($pull_index_wal_operation === "-") {
-                return [
-                    "path" => $remote_absolute_path,
-                    "delete" => true,
-                    "ctime" => 0,
-                    "size" => 0,
-                    "type" => null,
-                ];
-            }
-            if ($pull_index_wal_operation === "+") {
-                return [
-                    "path" => $remote_absolute_path,
-                    "delete" => false,
-                    "ctime" => (int) ($pull_index_wal_record["remote_path_ctime"] ?? 0),
-                    "size" => (int) ($pull_index_wal_record["remote_path_size"] ?? 0),
-                    "type" => (string) ($pull_index_wal_record["remote_path_type"] ?? "file"),
-                ];
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Reads one remote index update, keeping the last consecutive update for
-     * the same remote absolute path.
-     *
-     * @param mixed      $pull_index_wal_file_handle Open WAL handle.
-     * @param array|null $remote_index_update_lookahead Retained lookahead.
-     */
-    private function read_remote_index_update(
-        $pull_index_wal_file_handle,
-        ?array &$remote_index_update_lookahead = null
-    ): ?array {
-        if (!$pull_index_wal_file_handle) {
-            return null;
-        }
-        $current_remote_index_update =
-            $remote_index_update_lookahead
-            ?? $this->read_raw_remote_index_update(
-                $pull_index_wal_file_handle
-            );
-        $remote_index_update_lookahead = null;
-        if ($current_remote_index_update === null) {
-            return null;
-        }
-
-        while (true) {
-            $next_remote_index_update = $this->read_raw_remote_index_update(
-                $pull_index_wal_file_handle
-            );
-            if ($next_remote_index_update === null) {
-                return $current_remote_index_update;
-            }
-            if (
-                $next_remote_index_update["path"]
-                !== $current_remote_index_update["path"]
-            ) {
-                $remote_index_update_lookahead =
-                    $next_remote_index_update;
-                return $current_remote_index_update;
-            }
-            $current_remote_index_update = $next_remote_index_update;
-        }
     }
 
     /**
@@ -8800,7 +8225,7 @@ class ImportClient
         $path_prefix_found = false;
         while (($next_remote_index_json_line = fgets($next_remote_index_file_handle)) !== false) {
             try {
-                $next_remote_index_entry = $this->parse_index_line($next_remote_index_json_line);
+                $next_remote_index_entry = parse_index_line($next_remote_index_json_line);
             } catch (RuntimeException $e) {
                 continue;
             }
@@ -9474,12 +8899,12 @@ class ImportClient
             $file_changed = ($headers["x-file-changed"] ?? "0") === "1";
 
             if ($context->file_ctime && !$file_changed) {
-                $this->record_pulled_path(
+                $this->pull_index_journal->record_remote_upsert(
                     $path,
-                    $context->file_path,
                     $context->file_ctime,
                     $file_size,
                     "file",
+                    $context->file_path,
                 );
                 $this->files_pulled++; // Count completed files only
                 $this->clear_volatile_file($path);
@@ -9749,7 +9174,7 @@ class ImportClient
                 $this->audit_log("PRESERVE-LOCAL skip directory (exists): {$remote_absolute_path}", true);
                 $this->emit_skip_progress($remote_absolute_path);
                 if ($ctime > 0) {
-                    $this->upsert_remote_index_entry($remote_absolute_path, $ctime, 0, "dir");
+                    $this->pull_index_journal->record_remote_upsert($remote_absolute_path, $ctime, 0, "dir");
                 }
                 return;
             }
@@ -9757,7 +9182,7 @@ class ImportClient
                 $this->audit_log("PRESERVE-LOCAL skip directory (symlink in path): {$remote_absolute_path}", true);
                 $this->emit_skip_progress($remote_absolute_path);
                 if ($ctime > 0) {
-                    $this->upsert_remote_index_entry($remote_absolute_path, $ctime, 0, "dir");
+                    $this->pull_index_journal->record_remote_upsert($remote_absolute_path, $ctime, 0, "dir");
                 }
                 return;
             }
@@ -9788,12 +9213,12 @@ class ImportClient
         $this->audit_log("Directory: {$remote_absolute_path}", false);
 
         if ($ctime > 0) {
-            $this->record_pulled_path(
+            $this->pull_index_journal->record_remote_upsert(
                 $remote_absolute_path,
-                $local_absolute_path,
                 $ctime,
                 0,
-                "dir"
+                "dir",
+                $local_absolute_path
             );
         }
     }
@@ -9946,12 +9371,12 @@ class ImportClient
         $this->audit_log("Symlink: {$path} -> {$target_for_local}", false);
 
         if ($ctime > 0) {
-            $this->record_pulled_path(
+            $this->pull_index_journal->record_remote_upsert(
                 $path,
-                $local_absolute_path,
                 $ctime,
                 0,
-                "link"
+                "link",
+                $local_absolute_path
             );
         }
 
@@ -10009,7 +9434,7 @@ class ImportClient
             if (file_exists($local_absolute_path)) {
                 @unlink($local_absolute_path);
             }
-            $this->wal_append_remote_index_invalidation($path);
+            $this->pull_index_journal->record_remote_invalidation($path);
 
             if ($error_type === "file_changed") {
                 $this->record_volatile_file($path);
@@ -11489,9 +10914,9 @@ class ImportClient
         $this->shutdown_requested = true;
         $this->progress->clear_progress_line();
 
-        if (is_resource($this->pull_index_wal_handle)) {
+        if ($this->pull_index_journal->is_open()) {
             try {
-                $this->apply_pull_index_wal();
+                $this->pull_index_journal->apply_pending();
             } catch (Exception $e) {
                 $this->audit_log(
                     "Failed to apply the pull index WAL on shutdown: " .
